+++ scene_manager.h
@@ -18,7 +18,7 @@ public:
 	static void Draw();
 	static bool SetScene(Scene* pScene);
 	static Scene* GetScene();
-	static double GetProcessMS();
+	static float GetProcessMS();
 
 private:
 	static Scene* pScene_;
